fix delete image file error

+++ util/file.js
@@ -1,17 +1,24 @@
 const fs = require('fs');
 
 const deleteFile = (filePath) => {
-    let newFilePath;
-    if(filePath.charAt(0) === '/'){
-        newFilePath = filePath.substr(1);
-    } else{
-        newFilePath = filePath;
-    }
-    fs.unlink(newFilePath, (err) => {
-        if(err){
-            throw (err);
+    try {
+        if (fs.existsSync(path)) {
+            let newFilePath;
+            if(filePath.charAt(0) === '/'){
+                newFilePath = filePath.substr(1);
+            } else{
+                newFilePath = filePath;
+            }
+            fs.unlink(newFilePath, (err) => {
+                if(err){
+                    throw (err);
+                }
+            });
         }
-    });
+    } catch(err) {
+    console.error(err)
+    }
+    
 };
 
 exports.deleteFile = deleteFile;
